Update to support non-int64 IDs
Pleroma/Akkoma and GotoSocial use opaque IDs rather than `int64`s like
Mastodon which means that `madon` can't talk to either of those.

This (unfortunately big) commit updates everything that can be an ID to
`ActivityID` which is an alias for `string` - can't create a specific
type for it since there's more than a few places where they're
concatenated directly to strings for URLs, etc.  Which means it could
just as easily be a direct `string` type itself but I find that having
distinct types can often make the code more readable and understandable.